feat: switch store instances over to separate modules
