+++ src/utilities.js
@@ -1,5 +1,7 @@
 'use strict';
 
+import fetch from 'node-fetch';
+
 export const makeid = () => {
   let text = '';
   const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
@@ -18,8 +20,16 @@ export const wfcdn = (imgName) => `https://cdn.warframestat.us/img/${imgName}`;
 export const optimize = (img, size) =>
   `https://cdn.warframestat.us/${size ? `rs_${size}_fit,` : ''}o_webp,progressive_true/${img}`;
 
+export const get = async (url, opts) => {
+  try {
+    return fetch(url, opts).then((d) => d.json());
+    // eslint-disable-next-line no-empty
+  } catch (ignored) {}
+};
+
 export default {
   makeid,
   optimize,
   cdn,
+  get,
 };
